+++ skydome.cpp
@@ -181,7 +181,7 @@ bool CSkyDome::SetSunPosition( glm::vec3 const &Direction ) {
     }
 
     m_sundirection = Direction;
-	m_thetasun = std::acosf( m_sundirection.y );
+	m_thetasun = std::acos( m_sundirection.y );
 	m_phisun = std::atan2( m_sundirection.z, m_sundirection.x );
 
     return true;
